BackEnd/requirements.txt
Django==6.0.4
djangorestframework==3.17.1
django-cors-headers
pymysql==1.1.2